improvement(assetcard.tsx): stop deconstructing props needlessly
With the otherProps object being passed down to the Card component we
don't need to deconstruct them out of the this.props object only to set
them as JSX attributes on the Card component.

If we wanted to follow the pattern of setting them as JSX attributes
individually then we might as well have set all the attributes that the
Card component could potentially recieve to stay consistent.

I think that would be more work than necessary so we are going to pick a path
here and go with the attributes just being included in the otherProps
object being spread for the Card component.

When I carry on with further defining the Prop type for the AssetCard
component as being a union between the AssetCard props and the Card
props this will make things smoother.