improvement(assetcard.tsx): intersect AssetCard and Card propsTypes
In the AssetCard component folks had the ability to pass down props into the Card component via the
otherProps object. This commit makes that a little bit more obvious by intersecting the
AssetCardProps and CardProps type definition. It also would avoid TS errors if you tried to do that
sort of thing. Furthermore it also sets us up nice for the future when the CardProps type
potentially changes and instead of adding more types to the AssetCard we just get those for free.

#460